services/datahub/requirements.txt
# DataHub Service Dependencies

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1

# HTTP client for external APIs
httpx==0.25.2
aiohttp==3.9.1
python-binance==1.0.19

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
prometheus-client==0.19.0
apscheduler==3.10.4
pybreaker==1.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-socket==0.6.0
python-multipart==0.0.20
httpx==0.25.2

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

